# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: DomainsManager API
  version: 0.1.0
  summary: DomainsManager 后端 API 初版契约
  description: |
    本文档定义 FastAPI HTTP 接口。健康检查、本地认证、当前用户、用户域名、刷新任务、检查历史和基础管理员接口已实现。OAuth2 仅提供 Provider 配置状态；真实第三方授权和账号绑定仍待实现。
    所有时间均为 UTC RFC 3339 时间；除公开认证端点外均使用 Bearer Access Token。
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []
paths:
  /site/config:
    get:
      tags: [Site configuration]
      operationId: getPublicSiteConfig
      security: []
      parameters:
        - name: If-None-Match
          in: header
          required: false
          schema: { type: string }
      responses:
        '200':
          description: Public site presentation configuration
          headers:
            ETag: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PublicSiteConfig' }
        '304': { description: Configuration has not changed }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 短时有效的 Access Token。Refresh Token 不通过此方案传递。
...
  schemas:
...
    FooterLink:
      type: object
      additionalProperties: false
      required: [label, url]
      properties:
        label: { type: string, minLength: 1, maxLength: 80 }
        url: { type: string, minLength: 1, maxLength: 2048 }
    PublicSiteConfig:
      type: object
      additionalProperties: false
      required: [revision, site_name, site_logo, site_favicon, footer_links, footer_copyright, icp_number, police_record_number, custom_css, custom_javascript, head_html, body_end_html, analytics_code]
      properties:
        revision: { type: string }
        site_name: { type: string }
        site_logo: { type: string }
        site_favicon: { type: string }
        footer_links: { type: array, items: { $ref: '#/components/schemas/FooterLink' } }
        footer_copyright: { type: string }
        icp_number: { type: string }
        police_record_number: { type: string }
        custom_css: { type: string }
        custom_javascript: { type: string }
        head_html: { type: string }
        body_end_html: { type: string }
        analytics_code: { type: string }
...